requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client
google-api-python-client
aiogram
python-dotenv
httplib2
requests